feat: Allow relative URIs in the client
This is needed to allow Wasmer Edge to use the client for proxying
requests to Webassembly instances.

The upstream code requires absolute URIs, which are unexpected for HTTP 1.1.

This "hack" also allows relative URIs.

Since this change is purely additive, it should not negatively impact
any existing users of the client.